refactor: move BooleanEncoding into own module
